docker/back/requirements.txt
asgiref==3.6.0
dj-database-url==1.2.0
Django==4.1.6
django-environ==0.10.0
django-filter==22.1
djangorestframework==3.14.0
importlib-metadata==6.0.0
Markdown==3.4.1
pytz==2022.7.1
sqlparse==0.4.3
zipp==3.13.0
gunicorn==19.9.0
psycopg2==2.8.6